fix travis-install.

+++ .travis-install.sh
@@ -5,10 +5,13 @@ for p in `cat submodules`; do
   path=("${path[@]}" "./$p")
 done
 
-[ -e restriction ] && for p in `cat restriction`; do
-  path=("${path[@]}" "$p")
-done
-
 cabal --version
 cabal sandbox init
-cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}"
+
+if [ -e restriction ]; then
+  cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}" `cat restriction`
+  cabal install --force-reinstalls `cat restriction`
+else
+  cabal install --force-reinstalls --only-dependencies --enable-tests "${path[@]}"
+  cabal install --force-reinstalls 
+fi
